+++ convex/_generated/api.d.ts
@@ -92,6 +92,7 @@ export declare const components: {
         "internal",
         {
           entryId: string;
+          order: "desc" | "asc";
           paginationOpts: {
             cursor: string | null;
             endCursor?: string | null;
@@ -145,16 +146,6 @@ export declare const components: {
         {
           created: boolean;
           entryId: string;
-          replacedVersion: {
-            contentHash?: string;
-            entryId: string;
-            filterValues: Array<{ name: string; value: any }>;
-            importance: number;
-            key?: string;
-            metadata?: Record<string, any>;
-            status: "pending" | "ready" | "replaced";
-            title?: string;
-          } | null;
           status: "pending" | "ready" | "replaced";
         }
       >;
@@ -182,6 +173,24 @@ export declare const components: {
         { entryId: string; startOrder: number },
         null
       >;
+      deleteByKeyAsync: FunctionReference<
+        "mutation",
+        "internal",
+        { beforeVersion?: number; key: string; namespaceId: string },
+        null
+      >;
+      deleteByKeySync: FunctionReference<
+        "action",
+        "internal",
+        { key: string; namespaceId: string },
+        null
+      >;
+      deleteSync: FunctionReference<
+        "action",
+        "internal",
+        { entryId: string },
+        null
+      >;
       findByContentHash: FunctionReference<
         "query",
         "internal",
@@ -200,6 +209,7 @@ export declare const components: {
           importance: number;
           key?: string;
           metadata?: Record<string, any>;
+          replacedAt?: number;
           status: "pending" | "ready" | "replaced";
           title?: string;
         } | null
@@ -215,6 +225,7 @@ export declare const components: {
           importance: number;
           key?: string;
           metadata?: Record<string, any>;
+          replacedAt?: number;
           status: "pending" | "ready" | "replaced";
           title?: string;
         } | null
@@ -223,7 +234,7 @@ export declare const components: {
         "query",
         "internal",
         {
-          namespaceId: string;
+          namespaceId?: string;
           order?: "desc" | "asc";
           paginationOpts: {
             cursor: string | null;
@@ -245,6 +256,7 @@ export declare const components: {
             importance: number;
             key?: string;
             metadata?: Record<string, any>;
+            replacedAt?: number;
             status: "pending" | "ready" | "replaced";
             title?: string;
           }>;
@@ -257,13 +269,14 @@ export declare const components: {
         "internal",
         { entryId: string },
         {
-          replacedVersion: {
+          replacedEntry: {
             contentHash?: string;
             entryId: string;
             filterValues: Array<{ name: string; value: any }>;
             importance: number;
             key?: string;
             metadata?: Record<string, any>;
+            replacedAt?: number;
             status: "pending" | "ready" | "replaced";
             title?: string;
           } | null;
@@ -271,6 +284,29 @@ export declare const components: {
       >;
     };
     namespaces: {
+      deleteNamespace: FunctionReference<
+        "mutation",
+        "internal",
+        { namespaceId: string },
+        {
+          deletedNamespace: null | {
+            createdAt: number;
+            dimension: number;
+            filterNames: Array<string>;
+            modelId: string;
+            namespace: string;
+            namespaceId: string;
+            status: "pending" | "ready" | "replaced";
+            version: number;
+          };
+        }
+      >;
+      deleteNamespaceSync: FunctionReference<
+        "action",
+        "internal",
+        { namespaceId: string },
+        null
+      >;
       get: FunctionReference<
         "query",
         "internal",
@@ -335,6 +371,37 @@ export declare const components: {
           splitCursor?: string | null;
         }
       >;
+      listNamespaceVersions: FunctionReference<
+        "query",
+        "internal",
+        {
+          namespace: string;
+          paginationOpts: {
+            cursor: string | null;
+            endCursor?: string | null;
+            id?: number;
+            maximumBytesRead?: number;
+            maximumRowsRead?: number;
+            numItems: number;
+          };
+        },
+        {
+          continueCursor: string;
+          isDone: boolean;
+          page: Array<{
+            createdAt: number;
+            dimension: number;
+            filterNames: Array<string>;
+            modelId: string;
+            namespace: string;
+            namespaceId: string;
+            status: "pending" | "ready" | "replaced";
+            version: number;
+          }>;
+          pageStatus?: "SplitRecommended" | "SplitRequired" | null;
+          splitCursor?: string | null;
+        }
+      >;
       lookup: FunctionReference<
         "query",
         "internal",
@@ -351,7 +418,7 @@ export declare const components: {
         "internal",
         { namespaceId: string },
         {
-          replacedVersion: null | {
+          replacedNamespace: null | {
             createdAt: number;
             dimension: number;
             filterNames: Array<string>;
@@ -385,6 +452,7 @@ export declare const components: {
             importance: number;
             key?: string;
             metadata?: Record<string, any>;
+            replacedAt?: number;
             status: "pending" | "ready" | "replaced";
             title?: string;
           }>;